move sequencer block subset to data availability module
