+++ crates/astria-conductor/src/lib.rs
@@ -15,7 +15,6 @@ pub mod config;
 pub(crate) mod data_availability;
 pub(crate) mod executor;
 pub(crate) mod sequencer;
-pub(crate) mod types;
 
 pub use conductor::Conductor;
 pub use config::Config;
